requirements.txt
# Infrastructure dependencies
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Web dependencies
streamlit==1.38.0
requests==2.32.3

# GigaChat integration
gigachat==0.1.42
langchain==0.3.3
langchain-gigachat==0.3.12

# Additional utilities
tenacity==8.5.0
numpy==1.26.4
urllib3==2.2.3

# Development and testing
pytest==8.3.3
pytest-asyncio==0.24.0
